+++ modules/spring-web/src/main/kotlin/com/esprito/spring/web/editor/openapi/OpenApiUIEditor.kt
@@ -1,0 +1,9 @@
+package com.esprito.spring.web.editor.openapi
+
+import com.intellij.openapi.fileEditor.FileEditor
+import com.intellij.openapi.fileEditor.TextEditor
+import com.intellij.openapi.fileEditor.TextEditorWithPreview
+
+class OpenApiUIEditor(textEditor: TextEditor, preview: FileEditor) :
+    TextEditorWithPreview(textEditor, preview, "OpenAPI Preview Editor", Layout.SHOW_EDITOR_AND_PREVIEW) {
+}